Søylediagram som sammenligner fordeling (vha. stablede søyler) av valgt variabel

Søylediagram som sammenligner fordeling (vha. stablede søyler) av valgt variabel hos fra to ulike utvalg, f.eks. eget/resten, før/etter (Kan velge å ikke ha med sammenligninga) Søylediagram som viser andeler av valgt variabel: NevrNivaaInn' - Nevrologisk nivå ved innskrivelse, sml. eget, andre NevrNivaaUt' - Nevrologisk nivå ved utskrivelse, sml. eget, andre NevrNivaaInnUt' - sml. Nevrologisk nivå ved inn- og utskrivelse Hvilke kombinasjonsmuligheter har vi? Inn / Ut (Velge: bare eget/hele landet) Eget / resten av landet (Velge: Inn eller Ut)

NSFigAndelStabel(
  RegData,
  outfile = "",
  valgtVar,
  datoFra = "2010-01-01",
  datoTil = "3000-01-01",
  minald = 0,
  maxald = 130,
  erMann = "",
  traume = 99,
  preprosess = 1,
  enhetsUtvalg = enhetsUtvalg,
  reshID,
  hentData = 0,
  ...
)

Arguments

RegData

- ei dataramme med alle nødvendige variable fra registeret.

outfile

- navn på fil figuren skrives ned til.

valgtVar

- variabelen det skal genereres resultat for

datoFra

startdato, fra og med Eks. '2010-01-01' # min og max dato i utvalget vises alltid i figuren.

datoTil

sluttdato, fra og med Eks. '2013-05-25'

minald

- alder, fra og med

maxald

- alder, til og med

erMann

- kjønn, 1-menn, 0-kvinner, standard: ” (alt annet enn 0 og 1), dvs. begge

traume

- 'ja','nei', 'alle' standard: ikke valgt

preprosess

Preprosesser data 0: Nei 1: Ja (Standard)

enhetsUtvalg

- 1:eget sykehus, 0:hele landet (standard) Kun for valgtVar=='NevrNivaaInnUt'

reshID

- avdelingsid for egen avdeling, standard: 0-hele landet.

hentData

Gjør spørring mot database hvis data ikke er levert fra andre kilder. 0: Nei, RegData gis som input til funksjonen (Standard) 1: Ja