Funksjon som genererer en figur med gjennomsnitt/median

Funksjon som genererer en figur med gjennomsnitt/median for hvert sykehus og kan ta inn ulike numeriske variable. Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable enn sykehus.

NSFigGjsnGrVar(
  RegData,
  valgtVar,
  valgtMaal = "gjsn",
  grVar = "ShNavn",
  datoFra = "2010-01-01",
  datoTil = "2050-12-31",
  datoUt = 0,
  AIS = "",
  minald = 0,
  maxald = 130,
  erMann = "",
  traume = "alle",
  nivaaUt = 99,
  preprosess = 1,
  outfile = "",
  hentData = 0,
  ...
)

Arguments

RegData

- ei dataramme med alle nødvendige variable fra registeret.

valgtVar

- variabelen det skal genereres resultat for

valgtMaal

- 'med' = median. Alt annet gir gjennomsnitt

datoFra

startdato, fra og med Eks. '2010-01-01' # min og max dato i utvalget vises alltid i figuren.

datoTil

sluttdato, fra og med Eks. '2013-05-25'

datoUt

Om man skal velge inn eller ut-dato som grunnlag for datofiltrering

AIS

- AISgrad ved innleggelse alle(”), velge en eller flere fra 1:5, mappes til: A,B,C,D,E

minald

- alder, fra og med

maxald

- alder, til og med

erMann

- kjønn, 1-menn, 0-kvinner, standard: ” (alt annet enn 0 og 1), dvs. begge

traume

- 'ja','nei', 'alle' standard: ikke valgt

nivaaUt

- Nivå ved utreise, flervalgs 0:tetraplegi, 1:paraplegi, 2:C1-4, 3:C5-8, 9:ukjent

preprosess

Preprosesser data 0: Nei 1: Ja (Standard)

outfile

- navn på fil figuren skrives ned til.

hentData

Gjør spørring mot database hvis data ikke er levert fra andre kilder. 0: Nei, RegData gis som input til funksjonen (Standard) 1: Ja

Details

Argumentet valgtVar har følgende valgmuligheter:

Alder: Aldersfordeling, 15-årige grupper

DagerRehab: antall dager med rehabilitering [RehabDy]

DagerTilRehab: antall dager før rehabilitering [BeforeRehDy]

OpphTot: Lengde på opphold – totalt [HosptlDy]

Permisjon: Antall døgn ute av sykehus [OutOfHosptlDy]