Utvikling over tid for gjennomsnitt/median av valgt variabel

Figuren viser gjennomsnitt/median per år med konfidensintervall for valgt variabel. I bakgrunn vises konfidensintervall for det man har valt å sammenlikne med.

NSFigGjsnTid(
  RegData,
  valgtVar = "Alder",
  datoFra = "2011-01-01",
  datoTil = "3000-12-31",
  datoUt = 0,
  tidsenhet = "Aar",
  minald = 0,
  maxald = 110,
  erMann = "",
  reshID = 0,
  enhetsUtvalg = 0,
  valgtMaal = "gjsn",
  AIS = "",
  traume = "alle",
  nivaaUt = 99,
  preprosess = 1,
  outfile = "",
  hentData = 0,
  lagFigur = 1,
  ...
)

Arguments

RegData

- ei dataramme med alle nødvendige variable fra registeret.

valgtVar

Hvilken variabel som skal visualiseres

datoUt

Om man skal velge inn eller ut-dato som grunnlag for datofiltrering

reshID

- avdelingsid for egen avdeling, standard: 0-hele landet.

valgtMaal

'gjsn': gir middelverdi (standard) 'med': gir median

preprosess

Preprosesser data 0: Nei 1: Ja (Standard)

outfile

- navn på fil figuren skrives ned til.

hentData

Gjør spørring mot database hvis data ikke er levert fra andre kilder. 0: Nei, RegData gis som input til funksjonen (Standard) 1: Ja

Value

Linjediagram som viser utvikling over tid for valgt variabel

Details

Argumentet valgtVar har følgende valgmuligheter:

alder: Pasientens alders