Søylediagram som viser flere kvalitetsindikatorer

Funksjonen genererer en figur med ei gruppe kvalitetsindikatorer.

NSFigKvalInd(
  NSdata,
  outfile = "",
  valgtVar,
  datoFra = "2010-01-01",
  datoTil = "2050-01-01",
  AIS = "",
  minald = 0,
  maxald = 130,
  erMann = 99,
  traume = "",
  nivaaUt = 99,
  enhetsUtvalg = 1,
  reshID,
  hentData = 0,
  preprosess = 1
)

Arguments

NSdata

- Alle nødvendige variable fra hvert skjema.

outfile

- navn på fil figuren skrives ned til.

valgtVar

- variabelen det skal genereres resultat for

datoFra

startdato, fra og med Eks. '2010-01-01' # min og max dato i utvalget vises alltid i figuren.

datoTil

sluttdato, fra og med Eks. '2013-05-25'

AIS

- AISgrad ved innleggelse alle(''), velge en eller flere fra 1:5, mappes til: A,B,C,D,E

minald

- alder, fra og med

maxald

- alder, til og med

erMann

- kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge

traume

- 'ja','nei', 'alle' standard: ikke valgt

nivaaUt

- Nivå ved utreise, flervalgs 0:tetraplegi, 1:paraplegi, 2:C1-4, 3:C5-8, 9:ukjent

enhetsUtvalg

- 1:eget sykehus sml med resten, 2:eget sykehus, 0:hele landet (standard)

reshID

- avdelingsid for egen avdeling, standard: 0-hele landet.

hentData

Gjør spørring mot database hvis data ikke er levert fra andre kilder. 0: Nei, RegData gis som input til funksjonen (Standard) 1: Ja

preprosess

Preprosesser data 0: Nei 1: Ja (Standard)

figurtype

angir hvilken figurtype som skal lages: andeler, gjsnGrVar

Value

Søylediagram samling av kvalitetsindikatorer

Details

Argumentet valgtVar kan bare være "kvalInd".