Funksjon for å tilrettelegge variable for beregning.

Denne funksjonen gjør utvalg og tilrettelegger variable (gitt ved valgtVar) til videre bruk. Videre bruk kan eksempelvis være beregning av AggVerdier eller gjennomsnitt. Funksjonen gjør også filtreringer som å fjerne ugyldige verdier for den valgte variabelen, samt ta høyde for avhengigheter med andre variable. Det er også her man angir aksetekster og titler for den valgte variabelen. Her kan mye hentes til analysebok

NSVarTilrettelegg(RegData, valgtVar, grVar = "", figurtype = "andeler")

Arguments

RegData

- ei dataramme med alle nødvendige variable fra registeret.

valgtVar

- variabelen det skal genereres resultat for

figurtype

angir hvilken figurtype som skal lages: andeler, gjsnGrVar

Value

Definisjon av valgt variabel.

Details

Argumentet valgtVar har følgende valgmuligheter:

Alder: Aldersfordeling, 10-årige grupper

NB: valgtVar må komponeres slik at den henspeiler på hvilken tabell variabelen kommer fra. Dette for å senere kunne hente og koble dataene riktig. valgtSkjema <- substr(valgtVar,1,4) Prefixet til valgtVar skal være følgende:

Livs - LifeQuality,

Urin - UrinaryTractFunction

Tarm - BowelFunction

Kont - Control

Funk - ActivityAndParticipationPerformance

Tilf - ActivityAndParticipationSatisfaction

Argumentet enhetsUtvalg har følgende valgmuligheter:

0: Hele landet

1: Egen enhet mot resten av landet (Standard)

2: Egen enhet